pull out shared Service for all bot API services
